QoL updates
Changed output name pattern. Now it checks for the existence of previous run of the same input file and names the new one accordingly.
It also creates a log file containing a copy of config and new infos such as number of blocks and timings.

Made also some changes related to next update.